Fix: Wrong length variable used.
The wrong length was used to determine a part of the pre-encode length.
This uses the right variable fixing an off-by-x issue.

We also clean up a few of the "during development" prints, as well as
panic on failure now (it won't show the panic during build, but it'll be
clear to see in the generated janet.c).